+++ CHANGELOG.md
@@ -66,6 +66,7 @@
 - fix: fix value returned by `getFieldProps()` to be empty string instead of null when format option is not null
 - fix: export types from index to simplify import declarations
 - fix: do not mark form as validated when `validateFields()` was successful
+- fix: fix lost of modified state after using `appendListItem()`
 
 ## v5.6.0 (2025-01-09)